% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acrEularRA.R
\name{acrEularRA}
\alias{acrEularRA}
\title{Helper function for creating an acrEularRA class.}
\usage{
acrEularRA(
  ljc = numeric(),
  sjc = numeric(),
  duration = numeric(),
  onset = NULL,
  assessment = NULL,
  apr = character(),
  crp = numeric(),
  esr = numeric(),
  crp.uln = 10,
  esr.uln = 15,
  serology = character(),
  ccp = numeric(),
  rf = numeric(),
  ccp.uln = 10,
  rf.uln = 20
)
}
\arguments{
\item{ljc}{large joint count. Numeric between 0 and 10 of total
number of swollen and/or tender large joints.}

\item{sjc}{small joint count. Numeric between 0 and 18 of total
number of swollen and/or tender small joints.}

\item{duration}{numeric patient’s self-report on the maximum duration
(in days) of signs and symptoms of any joint that is clinically
involved at the time of assessment.}

\item{onset}{Date signs and symptoms started.}

\item{assessment}{Date of initial assessment.}

\item{apr}{character acute phase reactant levels. "Normal" or "Abnormal"}

\item{crp}{numeric of C-reactive protein test result.}

\item{esr}{numeric of erythrocyte sedimentation rate test result.}

\item{crp.uln}{numeric for upper limit of normal for the C-reactive protein test.}

\item{esr.uln}{numeric for upper limit of normal for the erythrocyte sedimentation
rate test.}

\item{serology}{character CCP and/or rheumatoid factor levels. "Negative",
"Low" positive or "High" positive.}

\item{ccp}{numeric of ccp test result.}

\item{rf}{numeric of rheumatoid factor test result.}

\item{ccp.uln}{numeric for upper limit of normal for the ccp test}

\item{rf.uln}{numeric for upper limit of normal for the RF test}
}
\value{
An acrEularRA object.
}
\description{
Creates an acrEular RA object from different parameters. Converts dates to
duration value and serology and acute phase reactant values to
classifications.
}
\examples{
 obj1 <- acrEularRA(ljc=8, sjc=12, duration=43, apr="Normal", serology="High")
 obj2 <- acrEularRA(ljc=8, sjc=12,
             onset=as.Date("2010-01-01"), assessment=as.Date("2010-02-13"),
             crp=5, esr=12, ccp=32, rf=71)

 all.equal(obj1, obj2)

}
