% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responseRA.R
\name{acrRA}
\alias{acrRA}
\title{Helper function for creating an acrRA class.}
\usage{
acrRA(
  sjc = numeric(),
  tjc = numeric(),
  ptGA = numeric(),
  ptPain = numeric(),
  phGA = numeric(),
  haq = numeric(),
  apr = numeric()
)
}
\arguments{
\item{sjc}{numeric swollen joint count. Numeric between 0 and 28 of total
number of swollen.}

\item{tjc}{numeric tender joint count. Numeric between 0 and 28 of total
number of twollen.}

\item{ptGA}{numeric patient global assessment score between 0 and 100.}

\item{ptPain}{numeric patient pain assessment score between 0 and 100.}

\item{phGA}{numeric physician global assessment score between 0 and 100.}

\item{haq}{numeric health assessment questionairre score between 0 and 3.}

\item{apr}{numeric acute phase reactant value. Either CRP or ESR}
}
\value{
An acrRA object.
}
\description{
Creates an acr RA object from different parameters used in calculating
ACR20/50/70 scores.
}
\examples{
 obj1 <- acrRA(sjc=8, tjc=12, ptGA=50, ptPain=35, phGA=60, haq=0.850, apr=15)
 obj2 <- acrRA(sjc=4, tjc=7, ptGA=20, ptPain=25, phGA=30, haq=0.350, apr=10)

 acrResponse(obj1, obj2)


}
