% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responseRA.R
\name{acrResponse}
\alias{acrResponse}
\title{Calculate ACR 20/50/70 response}
\usage{
acrResponse(acr1, acr2, na.rm = TRUE)
}
\arguments{
\item{acr1}{acrRA object with initial measurements.}

\item{acr2}{acrRA object with follow up measurements.}
}
\description{
Calculate ACR 20/50/70 response from SJC, TJC, patient global assessment,
patient pain VAS, physician global assessment, HAQ and
measurements.
}
\examples{

  acrResponse(acr1, acr2)

}
