% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acrEularRA.R
\name{aprClassification}
\alias{aprClassification}
\title{Calculate acute phase reactant classification from test scores and ULN.}
\usage{
aprClassification(crp, esr, crp.uln = 10, esr.uln = 15)
}
\arguments{
\item{crp}{numeric of C-reactive protein test result.}

\item{esr}{numeric of erythrocyte sedimentation rate test result.}

\item{crp.uln}{numeric for upper limit of normal for the C-reactive protein test.}

\item{esr.uln}{numeric for upper limit of normal for the erythrocyte sedimentation
rate test.}
}
\description{
Calculates acute phase reactant classification for given the C-reactive
protein and ESR test scores and the upper limit of normal.
}
\examples{
aprClassification(crp=9, esr=16, crp.uln=10, esr.uln=15)

}
