% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acrEularRA.R
\name{aprScore}
\alias{aprScore}
\title{Calculate acute phase reactant component score}
\usage{
aprScore(object)
}
\arguments{
\item{object}{acrEularRA object}
}
\description{
Calculate acute phase reactant component score. Converts acute phase
reactant status to a numeric score
}
\examples{
acreular <- new_acrEularRA(ljc=3,sjc=4,duration=60,apr="Abnormal",serology="High")
aprScore(acreular)

}
