% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{datesToDuration}
\alias{datesToDuration}
\title{Convert two dates to a duration in days.}
\usage{
datesToDuration(onset, assessment)
}
\arguments{
\item{onset}{date}

\item{assessment}{date}
}
\description{
Converts onset and assessment dates to a duration in days.
}
\examples{
onset <- as.Date("2010-01-01")
assessment <- as.Date("2010-02-13")

datesToDuration(onset, assessment)

}
