% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responseRA.R
\name{eularResponse}
\alias{eularResponse}
\title{Calculate EULAR response}
\usage{
eularResponse(das1, das2)
}
\arguments{
\item{numeric}{das1 Initial DAS28-ESR/CRP measurement(s)}

\item{numeric}{das2 Follow up DAS28-ESR/CRP measurement(s)}
}
\description{
Calculate EULAR response from two DAS28-ESR or DAS28-CRP
measurements.
}
\examples{
  eularResponse(5.31, 1.3)

  baseline <- c(5.24, 3.6, 1.2)
  followup <- c(1.30, 3.3, 1.8)

  eularResponse(baseline, followup)

}
