% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{is.Date}
\alias{is.Date}
\title{Test for objects of type "Date"}
\usage{
is.Date(x)
}
\arguments{
\item{x}{object to be tested.}
}
\description{
Checks if an object inherits class Date. Used prior to calculating durations
from two dates.
}
\examples{
date <- as.Date("2010-01-01")

is.Date(date)

}
