% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acrEularRA.R
\name{jointScore}
\alias{jointScore}
\title{Calculate joint component score}
\usage{
jointScore(object)
}
\arguments{
\item{object}{acrEularRA object}
}
\description{
Calculate joint component score. Converts patients swollen/tender joint
counts to a numeric score.
}
\examples{
acreular <- new_acrEularRA(ljc=3,sjc=4,duration=60,apr="Abnormal",serology="High")
jointScore(acreular)

}
