% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acrEularRA.R
\name{new_acrEularRA}
\alias{new_acrEularRA}
\title{Create a new acrEularRA class}
\usage{
new_acrEularRA(
  ljc = numeric(),
  sjc = numeric(),
  duration = numeric(),
  apr = character(),
  serology = character()
)
}
\arguments{
\item{ljc}{numeric large joint count. Number of swollen and/or tender
large joints.}

\item{sjc}{numeric small joint count. Number of swollen and/or tender
small joints.}

\item{duration}{numeric patient’s self-report on the maximum duration
(in days) of signs and symptoms of any joint that is clinically
involved at the time of assessment.}

\item{apr}{character acute phase reactant levels. "Normal" or "Abnormal"}

\item{serology}{character CCP and/or rheumatoid factor levels. "Negative",
"Low" positive or "High" positive.}
}
\value{
An acrEularRA object.
}
\description{
\code{new_acrEularRA} returns an acrEularRA object.
}
\examples{
 obj <- new_acrEularRA(ljc=8, sjc=12, duration=43, apr="Normal", serology="High")

}
