% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responseRA.R
\name{new_acrRA}
\alias{new_acrRA}
\title{Create a new acrRA class}
\usage{
new_acrRA(sjc = numeric(), tjc = numeric(), ptGA = numeric(),
  ptPain = numeric(), phGA = 60, haq = numeric(), apr = numeric())
}
\arguments{
\item{sjc}{numeric swollen joint count.}

\item{tjc}{numeric tender joint count.}

\item{ptGA}{numeric patient global assessment score.}

\item{ptPain}{numeric patient pain assessment score.}

\item{phGA}{numeric physician global assessment score.}

\item{haq}{numeric health assessment questionairre score.}

\item{apr}{numeric acute phase reactant value.}
}
\value{
An acrRA object.
}
\description{
\code{new_acrRA} returns an acrRA object.
}
\examples{
 obj <- new_acrRA(sjc=8, tjc=12, ptGA=50, ptPain=29, phGA=60, haq=0.875, apr=15)

}
