% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acrEularRA.R
\name{serologyClassification}
\alias{serologyClassification}
\title{Calculate serology classification from test scores and ULN}
\usage{
serologyClassification(ccp, rf, ccp.uln = 10, rf.uln = 20)
}
\arguments{
\item{score}{numeric of serology test result}

\item{uln}{numeric for upper limit of normal for the serology test}
}
\description{
Calculates serology classification for CCP and/or rheumatoid factor given
the test scores and the upper limit of normal..
}
\examples{
serologyClassification(ccp=9, rf=21, ccp.uln=10, rf.uln=20)

}
