% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acrEularRA.R
\name{serologyScore}
\alias{serologyScore}
\title{Calculate serology component score}
\usage{
serologyScore(object)
}
\arguments{
\item{object}{acrEularRA object}
}
\description{
Calculate joint component score. Converts patients serology status to
a numeric score.
}
\examples{
acreular <- new_acrEularRA(ljc=3,sjc=4,duration=60,apr="Abnormal",serology="High")
serologyScore(acreular)

}
