% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{shiny_acreular}
\alias{shiny_acreular}
\title{Launch shiny acreular interface}
\usage{
shiny_acreular(display.mode = "normal")
}
\arguments{
\item{display.mode}{The display mode to be passed to \link[shiny]{runApp}}
}
\description{
\code{shiny_acreular} launches a shiny interface for browser based ACR/EULAR calculations.
}
\examples{
\dontrun{
shiny_acreular()
shiny_acreular(display.mode="normal")
}
}
