% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acrEularRA.R
\name{acrEularRAScore}
\alias{acrEularRAScore}
\title{Calculate ACR/EULAR 2010 RA score}
\usage{
acrEularRAScore(object)
}
\arguments{
\item{object}{acrEularRA object}
}
\description{
Calculates ACR/EULAR 2010 RA score from the individual components.
}
\examples{
acreular <- new_acrEularRA(ljc=3,sjc=4,duration=60,apr="Abnormal",serology="High")
acrEularRAScore(acreular)

}
