% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responseRA.R
\name{acrResponse}
\alias{acrResponse}
\title{Calculate ACR 20/50/70 response}
\usage{
acrResponse(tjc1, tjc2, sjc1, sjc2, ptGA1, ptGA2, ptPain1, ptPain2, phGA1,
  phGA2, haq1, haq2, apr1, apr2)
}
\arguments{
\item{tjc1}{numeric initial tender joint count.}

\item{sjc1}{numeric initial swollen joint count.}

\item{sjc2}{numeric follow up swollen joint count.}

\item{ptGA1}{numeric initial patient global assessment score.}

\item{ptGA2}{numeric follow up patient global assessment score.}

\item{ptPain1}{numeric initial patient pain assessment score.}

\item{ptPain2}{numeric initial patient pain assessment score.}

\item{phGA1}{numeric initial physician global assessment score.}

\item{phGA2}{numeric follow up physician global assessment score.}

\item{haq1}{numeric initial health assessment questionairre score.}

\item{haq2}{numeric follow health assessment questionairre score.}

\item{apr1}{numeric initial acute phase reactant value.}

\item{apr2}{numeric follow up acute phase reactant value.}

\item{tjc1}{numeric follow up tender joint count.}
}
\description{
Calculate ACR 20/50/70 response from SJC, TJC, patient global assessment,
patient pain VAS, physician global assessment, HAQ and
measurements.
}
\examples{

  acrResponse(tjc1=6, tjc2=1, sjc1=6, sjc2=1,
              ptGA1=17, ptGA2=1, ptPain1=11, ptPain2=1,
              phGA1=60, phGA2=5, haq1=5, haq2=3, apr1=24, apr2=5)

}
