% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responseRA.R
\name{acrResponse}
\alias{acrResponse}
\title{Calculate ACR 20/50/70 response}
\usage{
acrResponse(acr1, acr2, na.rm = TRUE)
}
\arguments{
\item{acr1}{acrRA object with initial measurements.}

\item{acr2}{acrRA object with follow up measurements.}

\item{na.rm}{logical. Should missing values be removed.}
}
\description{
Calculate ACR 20/50/70 response from SJC, TJC, patient global assessment,
patient pain VAS, physician global assessment, HAQ and
measurements.
}
\examples{
  acr1 <- acrRA(sjc=8, tjc=12, ptGA=50, ptPain=35, phGA=60, haq=0.850, apr=15)
  acr2 <- acrRA(sjc=4, tjc=7, ptGA=20, ptPain=25, phGA=30, haq=0.350, apr=10)

  acrResponse(acr1, acr2)

}
