% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{is.Date}
\alias{is.Date}
\title{Checks if object is of class Date.}
\usage{
is.Date(x)
}
\arguments{
\item{date}{}
}
\description{
Checks if object is of class Date.
}
\examples{
date <- as.Date("2010-01-01")

is.Date(date)

}
